refactored App.js and added standalone Calculator component
