--- Project-II-Calculator/calculator/src/App.css
@@ -1,49 +1,0 @@
-.app {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  align-items: flex-start;
-  border: 2px solid lightgray;
-  width: 460px;
-  height: 600px;
-  margin: 40px auto;
-}
-
-.app .buttons {
-  display: flex;
-}
-
-.app .buttons .left {
-  display: flex;
-  flex-direction: column;
-  width: 80%;
-  height: 498px;
-}
-
-.app .buttons .clear {
-  width: 100%;
-  height: 20%;
-}
-
-.app .buttons .clear button {
-  font-weight: 300;
-}
-
-.app .buttons .number-buttons {
-  display: flex;
-  flex-wrap: wrap;
-  width: 100%;
-  height: 60%;
-}
-
-.app .buttons .zero {
-  width: 100%;
-  height: 20%;
-}
-
-.app .buttons .calc-button {
-  display: flex;
-  flex-direction: column;
-  width: 28%;
-  height: 498px;
-}
