+++ Project-II-Calculator/calculator/src/components/CalculatorComponent/Calculator.css
@@ -1,0 +1,49 @@
+.calculator {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: flex-start;
+  align-items: flex-start;
+  border: 2px solid lightgray;
+  width: 460px;
+  height: 600px;
+  margin: 40px auto;
+}
+
+.calculator .buttons {
+  display: flex;
+}
+
+.calculator .buttons .left {
+  display: flex;
+  flex-direction: column;
+  width: 80%;
+  height: 498px;
+}
+
+.calculator .buttons .clear {
+  width: 100%;
+  height: 20%;
+}
+
+.calculator .buttons .clear button {
+  font-weight: 300;
+}
+
+.calculator .buttons .number-buttons {
+  display: flex;
+  flex-wrap: wrap;
+  width: 100%;
+  height: 60%;
+}
+
+.calculator .buttons .zero {
+  width: 100%;
+  height: 20%;
+}
+
+.calculator .buttons .calc-button {
+  display: flex;
+  flex-direction: column;
+  width: 28%;
+  height: 498px;
+}
